feat: 测试用例改造为ts
