+++ package.json
@@ -14,9 +14,7 @@
     "prepublishOnly": "npm test && npm run build",
     "tar": "npm pack && cp -r ./*.tgz examples/san-loader.tgz && rimraf ./*.tgz",
     "local": "npm run build && cp -r dist/* examples/complex/node_modules/san-loader/dist",
-    "test": "run-p test:*",
-    "test:unit": "jest test/unit --no-cache",
-    "test:e2e": "jest test/e2e --no-cache"
+    "test": "npm run build && jest --coverage"
   },
   "keywords": [
     "san",
@@ -40,14 +38,16 @@
   ],
   "license": "MIT",
   "devDependencies": {
+    "@types/jest": "^27.4.0",
     "@types/loader-utils": "^2.0.3",
     "@types/qs": "^6.9.7",
-    "jest": "^27.3.1",
+    "jest": "^27.5.1",
     "npm-run-all": "^4.1.5",
     "prettier": "^2.4.1",
     "rimraf": "^3.0.2",
     "san-anode-utils": "^3.10.0",
-    "typescript": "^4.4.4",
+    "ts-jest": "^27.1.3",
+    "typescript": "^4.5.5",
     "webpack5": "npm:webpack@^5.64.0"
   },
   "dependencies": {
@@ -64,8 +64,5 @@
   "prettier": {
     "singleQuote": true,
     "trailingComma": "es5"
-  },
-  "jest": {
-    "testTimeout": 15000
   }
 }
