--- test/e2e/complex.spec.js
@@ -1,16 +1,0 @@
-const build = require('../webpack-build');
-
-describe('complex', () => {
-  let result;
-
-  beforeAll(async () => {
-    result = await build('complex');
-  });
-  afterAll(async () => {
-    await new Promise((resolve) => setTimeout(() => resolve(), 500));
-  });
-
-  it('🌟 should compile <template> with webpack4', () => {
-    expect(result).toBeTruthy();
-  });
-});
